Bugfix: make mmalloc build in x64
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@7727 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ src/xbt/mmalloc/mmap-sup.c
@@ -194,7 +194,7 @@ mmalloc_findbase (int size)
 	 to signal an error return, and besides, it is useful to
 	 catch NULL pointers if it is unmapped.  Instead start
 	 at the next page boundary. */
-      base = (void*) getpagesize ();
+      base = (void*)(long) getpagesize ();
     }
   else if (base == (void*) -1)
     {
